feat: "limit" query parameter for GET request
respond to the GET HTTP request with a :limit query parameter which stands for the maximum products wanted on the result set.
findLimitedProducts(int limit) method was added in the ProductRepo class. It is a Jpa method, able to excecute the query inside the @Query notion. This method a returns a list of products ordered in the ascending order as specified in the query. If no product is present in the database, the returned list is empty.
In this case, a RessourceNotFound Exception is thrown ( 404 NOT FOUND ).
Also, the service layer makes sure of the input value - of limit. limit value must be strictly positive.
If limit equals 0, we end up with an empty set as query result, which causes RessourceNotFound Exception.
If limit value is negative, it is clear that is a bad request.
limit = 0 case is considered as Bad Request.