fix: http bad request responses
make the api able to respond with 400 bad request http status code when json keys are bad.
this solution helps the client to know what is wrong with its data. So instead of responding with a 404 not found or 500 internal server error ( when a certain value is null ) now the client know that there is problem with its data.
The api tells the client if the id key wrong and provides him with the correct key.
The same thing apply to the name attribute.
The quantity attribute can be null, so if wrong key is provided, quantity is set to 0