feat: get products ordered and limit results
Make the api respond to "order" query parameter to determine the order strategy and the "limit" query parameter, which is an interegs that stands for the number of products wanted in the result set.
the order accept two values: asc for ascendent order.
    and desc for descendent order
Any value different then these two values is not accepted and result a 422 status code.

limit accepts any positive integer. if the positive integer is greater then the number of entries in the database, the hole dataset will be returned.
of the limit value is zero or negative, it is considered as bad requests.